requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
Django==5.0.4
django-mdeditor==0.1.20
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.3